Refactored game.rb into filer and difficulty modules. With re-wrriten tests.
