+++ lib/mathangman/game.rb
@@ -1,11 +1,14 @@
 require_relative "utility"
-require "time"
-require "fileutils"
+require_relative "difficulty"
+require_relative "filer"
 
 module Mathangman
+
   class Game
 
     include Utility
+    include Difficulty
+    include Filer
 
     attr_reader :display
     attr_accessor :secret_word, :name, :folder, :restart, :files, :word, :len, :wrongs_num, :disp_word, :guess
@@ -16,23 +19,8 @@ module Mathangman
       @display = disp unless disp.nil?
     end
 
-    def diff_level(diff)
-      @diff = diff
-      if diff == "7"
-        @wrongs_num = @len / 2 + @guess_bonus
-        return unusable = false if @len > 4 && @len < 13
-      elsif diff == "8"
-        @wrongs_num = @len / 2 + ( @guess_bonus - 1 )
-        return unusable = false if @len > 6 && @len < 13
-      elsif diff == "9"
-        @wrongs_num = @len / 2
-        return unusable = false if @len > 9 && @len < 13
-      end
-      true
-    end
-
     def rand_word(diff = nil)
-      dict = Mathangman::Utility.check_source "./5desk.txt"
+      dict = check_source "./5desk.txt"
       unusable = true
       while unusable
         @secret_word = dict.sample.chomp.downcase
@@ -52,17 +40,17 @@ module Mathangman
       @guess = gets.chomp.downcase
     end
 
-    def guesses
+    def guesses(obj = nil)
       puts @display.msg "Enter an alphabet guess for the #{@len} letter word."
       input_from_user
       if @guess == "*"
-        Mathangman::Utility::quitter(self)
+        quitter(self)
       end
       process
     end
 
     def process
-      if Mathangman::Utility.is_alpha? @guess
+      if is_alpha? @guess
         act_on_guess
       else
         puts @display.msg "Invalid input. Only alphabets are allowed."
@@ -118,8 +106,8 @@ module Mathangman
     def player_choice(choice)
       if supported_actions.include? choice
         send(supported_actions[choice])
-      elsif "*"
-        Mathangman::Utility.quitter("force")
+      elsif choice == "*"
+        quitter("force")
       else
         puts @display.invalid_entry
         show_disp_menu
@@ -149,21 +137,7 @@ module Mathangman
     end
 
     def check_validity
-      true if Mathangman::Utility.is_alpha?(@name)
-    end
-
-    def check_difficulty
-      diff = gets.chomp
-      levels = ["7", "8", "9"]
-      if levels.include? diff
-        first_guess(diff)
-      elsif diff == "*"
-        Mathangman::Utility::quitter("pre_game")
-      else
-        puts @display.invalid_entry
-        puts @display.difficulty
-        check_difficulty
-      end
+      true if is_alpha? @name
     end
 
     def show_disp_menu
