+++ saved_games/1444150495.txt
@@ -1,0 +1,5 @@
+7
+-o----
+solidi
+3
+6
