+++ saved_games/ola/1444150684.txt
@@ -1,0 +1,5 @@
+9
+-o---o-o----
+toxicologist
+5
+12
