+++ saved_games/olaide/1444150628.txt
@@ -1,0 +1,5 @@
+9
+-a----------
+satisfyingly
+6
+12
