+++ spec/game_spec.rb
@@ -5,7 +5,7 @@ describe Mathangman::Game do
 
   before :each  do
     @disp = Mathangman::Display.new
-     @game = Mathangman::Game.new(@disp)
+    @game = Mathangman::Game.new(@disp)
   end
 
   describe "#new" do
@@ -18,15 +18,6 @@ describe Mathangman::Game do
     end
   end
 
-  describe "#diff_level" do
-    it "returns true " do
-      expect(@game.diff_level(4)).to be true
-    end
-
-    it "returns true " do
-      expect(@game.diff_level("r")).to be true
-    end
-  end
 
   describe "#show_letter" do
 
@@ -68,30 +59,12 @@ describe Mathangman::Game do
     end
   end
 
-  describe "#create_dir" do
-    it "returns nil " do
-      @game.name = "olaide"
-      expect(@game.create_dir).to be nil
-    end
-  end
 
-  describe "#del_dir" do
-    it "returns nil " do
-      expect(@game.del_dir).to be nil
-    end
-  end
-
-  describe "#folder_not_exist?" do
-    it "returns nil " do
-      expect(@game.folder_not_exist?).to be_falsy
-      expect{ @game.folder_not_exist? "arg" }.to raise_error ArgumentError
-    end
-  end
 
   describe "#check_validity" do
     it "returns true " do
       # require "pry";   binding.pry
-      allow(Mathangman::Utility).to receive(:is_alpha?).and_return(true)
+      allow(@game).to receive(:is_alpha?).and_return(true)
       expect(@game.check_validity).to be true
       expect{ @game.check_validity "arg" }.to raise_error ArgumentError
     end
@@ -99,52 +72,8 @@ describe Mathangman::Game do
 
   describe "#check_validity" do
     it "returns false " do
-      allow(Mathangman::Utility).to receive(:is_alpha?).and_return(false)
+      allow(@game).to receive(:is_alpha?).and_return(false)
       expect(@game.check_validity).to be_falsy
-    end
-  end
-
-  describe "#check_difficulty" do
-    it "returns falsy " do
-      allow(@game).to receive(:quitter).and_return(nil)
-      allow(@game).to receive(:first_guess).and_return(nil)
-      allow(@game).to receive(:gets).and_return("7")
-      expect(@game.check_difficulty).to be_falsy
-    end
-  end
-
-  describe "#load_file" do
-    it "returns falsy " do
-      allow(@game).to receive(:guesses).and_return(nil)
-      allow(@disp).to receive(:msg).and_return(nil)
-      allow(@game).to receive(:restore_state).and_return(nil)
-      allow(@game).to receive(:gets).and_return("0")
-      @game.files =["test_games.txt"]
-      expect(@game.load_file).to be_falsy
-    end
-  end
-
-  describe "#restore_state" do
-    it "returns nil " do
-      expect(@game.restore_state(["info","e","w","2","e"])).to be nil
-    end
-  end
-
-  describe "#restore_state" do
-    it "throws inappropraite ArgumentError" do
-      expect{ @game.restore_state }.to raise_error ArgumentError
-    end
-  end
-
-  describe "#save_state" do
-    it "returns nil if called with appropriate character " do
-      expect(@game.save_state("loc")).to be nil
-    end
-  end
-
-  describe "#save_state" do
-    it "throws inappropraite ArgumentError " do
-      expect{ @game.save_state }.to raise_error ArgumentError
     end
   end
 
@@ -182,7 +111,7 @@ describe Mathangman::Game do
   describe "#player_choice" do
 
     it "gets player choice" do
-      allow(@game).to receive(:quitter).and_return(true)
+      # allow(@game).to receive(:quitter).and_return(true)
       expect{@game.player_choice '*'}.to raise_error SystemExit
     end
 
@@ -248,33 +177,7 @@ describe Mathangman::Game do
     end
   end
 
-  describe "#get_folder_items" do
-    it "returns a Hash" do
-      @game.folder
-      expect(@game.get_folder_items).to be_a Hash
-    end
-  end
 
-  describe "#load_game" do
-    it "returns nil" do
-      allow(@game).to receive(:puts).and_return(nil)
-      allow(@game).to receive(:gets).and_return("base")
-      allow(@game).to receive(:folder_not_exist?).and_return(true)
-      allow(@game).to receive(:load_file).and_return(nil)
-      allow(@game).to receive(:show_disp_menu).and_return(nil)
-      allow(@disp).to receive(:msg).and_return(nil)
-      @game.folder
-      expect(@game.load_game).to be nil
-    end
-  end
-
-  describe "#save_game" do
-    it "returns nil" do
-      allow(@game).to receive(:check_repeated).and_return(String)
-      allow(@game).to receive(:save_state).and_return(nil)
-      expect(@game.save_game).to be nil
-    end
-  end
 
   describe "#act_on_guess" do
     it "returns nil" do
@@ -296,7 +199,7 @@ describe Mathangman::Game do
 
   describe "#process" do
     it "returns nil" do
-      allow(Mathangman::Utility).to receive(:is_alpha?).and_return(true)
+      allow(@game).to receive(:is_alpha?).and_return(true)
       allow(@game).to receive(:act_on_guess).and_return(nil)
       allow(@game).to receive(:puts).and_return(nil)
       allow(@game).to receive(:guesses).and_return(nil)
@@ -305,7 +208,7 @@ describe Mathangman::Game do
     end
 
     it "returns nil" do
-      allow(Mathangman::Utility).to receive(:is_alpha?).and_return(false)
+      allow(@game).to receive(:is_alpha?).and_return(false)
       allow(@game).to receive(:puts).and_return(nil)
       allow(@game).to receive(:guesses).and_return(nil)
       @game.wrongs_num
